==> 3-SummaryResults.Rmd <==
---
title: "Australia HBV Model Summary Results"
author: "Richard T. Gray"
date: Latest version - `r format(Sys.Date(), format="%B %d %Y")`
output: 
  word_document:
    pandoc_args: --output="docs/Summary_results.docx"
    reference_docx: docs/mystyles.docx
---

<!---This Rmarkdown scipt generates the summary figures and outputs from
the results generated using the ECDC HIV Modelling Tool. --->

```{r knitr_options, include=FALSE} 
knitr::opts_chunk$set(echo = FALSE, 
                      warning = FALSE, 
                      message = FALSE, 
                      include = FALSE) 
```

This document presents the summary results for the baseline scenario in
the Australian HBV model. 

```{r Initialize script}
# Clear workspace and plots
rm(list = ls()) 
graphics.off()

# Setup directories after setting working directory to source file 
# directory (so it is easy to move code around)
basePath <- getwd()
Rcode <- file.path(basePath, "code")
projectFolder <- file.path(basePath, "projects")

# Load useful libraries
source(file.path(Rcode, "LoadLibrary.R"))
source(file.path(Rcode, "DataLibraries.R"))
source(file.path(Rcode, "PlotOptions.R"))
source(file.path(Rcode, "PlotFunctions.R"))
LoadLibrary(cowplot)

# Wrapper function so we don't have to keep entering everything in
# ggsave. Plot defaults are put in the figure defaults
SaveFigure <- function(folder, filename, figure, 
                       format = ".png", height = 10, 
                       units = "cm", ...) {
  save_plot(file.path(folder, paste0(filename, format)), 
         figure, 
         base_height = height, units = units, ...)
}

```

```{r User sepcifications}
# Specify project to run
project <- "Australia_HBV_estimates-BM"

# Plotting options
displayPlots <- TRUE
savePlots <- TRUE

startYear <- 2000
endYear <- 2030

# Results time
resultsTime <- "base" # or enter the date and time string

```

```{r Load project results}
# This loads all the project parameters

projectFile <- file.path(projectFolder, project, paste0(project, ".rda"))
load(projectFile)

# This loads all the project results
resultsFile <- file.path(projectFolder, project, "results", 
                         paste0("results_", resultsTime), 
                          paste0("results_", resultsTime, ".rda"))

load(resultsFile)
```

```{r Generate plots}

popLabels <- c("Age < 4", "Age 5 to 14", "Age 15 to 44", "Age > 45")
stateLabels <- c("Susceptible", "Acute", "Chronic", "cleared", "immune")

# Population sizes ------------------------------------------------------
totalPop <- popResults(pg, bestResults, paramResults,
                       populations = "all", states = "all",
                       range = TRUE) 

totalPopPlot <- indicatorPlot(totalPop, ylabel = "Population size",
                              xlimits = c(startYear, endYear, 5))


# Plot of population size in every group 
popSizesAge <- popResults(pg, bestResults, paramResults,
                          states = "all", range = TRUE) %>%
  FactorPop(popLabels)

agePopPlot <- indicatorPlot(popSizesAge, ylabel = "Population size",
                              xlimits = c(startYear, endYear, 10),
                            facetPlot = "population", range = TRUE)

# Prevalence --------------------------------------------------

totalNumInfected <- popResults(pg, bestResults, paramResults,
                          states = c("a", "ch"), 
                          range = TRUE, populations = "all")

ageNumInfected <- popResults(pg, bestResults, paramResults,
                          states = c("a", "ch"), 
                          range = TRUE) %>%
  FactorPop(popLabels)

totalInfectedPlot <- indicatorPlot(totalNumInfected, 
                                   ylabel = "Total number infected",
                                   xlimits = c(startYear, endYear, 5),
                                   range = TRUE)

popInfectedPlot <- indicatorPlot(ageNumInfected, 
                                   ylabel = "Number infected",
                                   xlimits = c(startYear, endYear, 10),
                                   range = TRUE,
                                   facetPlot = "population")

# Total prevalance
tempInfected <- popResults(pg, bestResults, paramResults,
  states = c("a", "ch"), 
  range = FALSE, populations = "all") %>% 
  select(-year)

tempTotal <- popResults(pg, bestResults, paramResults,
                       populations = "all", states = "all",
                       range = FALSE) %>% select(-year)

tempPrev <- (100 * tempInfected / tempTotal) %>%
  mutate(year = totalPop$year) %>%
  select(year, everything()) %>%
  tbl_df()

totalPrev <- tempPrev %>%
  gather("sim", "popsize", 2:(ncol(tempInfected) + 1)) %>%
  group_by(year) %>%
  summarise(min = min(popsize),
            max = max(popsize)) %>%
  ungroup() %>%
  mutate(best = tempPrev$best) %>%
  select(year, best, min, max)

totalPrevPlot <- indicatorPlot(totalPrev, 
                               ylabel = "Overall prevalence (%)",
                               xlimits = c(startYear, endYear, 5),
                               range = TRUE)

# Pop prevalence
tempInfected <- popResults(pg, bestResults, paramResults,
                          states = c("a", "ch"), 
                          range = FALSE) %>% select(-year, -population)

tempTotal <- popResults(pg, bestResults, paramResults, states = "all",
                       range = FALSE) %>% select(-year, -population)

tempPrev <- (100 * tempInfected / tempTotal) %>%
  mutate(year = popSizesAge$year,
         population = popSizesAge$population) %>%
  select(year, population, everything()) %>%
  tbl_df()

popPrev <- tempPrev %>%
  gather("sim", "popsize", 3:ncol(tempPrev)) %>%
  group_by(year, population) %>%
  summarise(min = min(popsize),
            max = max(popsize)) %>%
  ungroup() %>%
  mutate(best = tempPrev$best) %>%
  select(year, population, best, min, max) %>%
  FactorPop(popLabels)

popPrevPlot <- indicatorPlot(popPrev, 
                             ylabel = "Population prevalence (%)",
                             xlimits = c(startYear, endYear, 10),
                             range = TRUE, facetPlot = "population")


# Plot of population size in every group- first sort out level names for
# plotting. Essentially prevalence by age and state
plotPops <- popResults(pg, bestResults, paramResults, range = TRUE) %>%
  FactorPop(popLabels) %>%
  FactorState(stateLabels)

allPopsPlot <- indicatorPlot(plotPops, 
                             ylabel = "Population size",
                             xlimits = c(startYear, endYear, 15),
                             range = TRUE, facetPlot = "population",
                             groupPlot = "state")

allStatesPlot <- indicatorPlot(plotPops, 
                             ylabel = "Population size",
                             xlimits = c(startYear, endYear, 15),
                             range = TRUE, facetPlot = "state",
                             groupPlot = "population")

# New infections -------------------------------------------

# Annual number of new infections
totalInfections <- indicatorResults(pg, bestResults, "newInfections",
                                    paramResults,
                                    range = TRUE,
                                    populations = "all", 
                                    annual = "sum")

popInfections <- indicatorResults(pg, bestResults, "newInfections",
                                  paramResults,
                                  range = TRUE,
                                  annual = "sum") %>%
  FactorPop(popLabels)

totalNewInfectionsPlot <- indicatorPlot(totalInfections, 
                                   ylabel = "New infections",
                                   xlimits = c(startYear, endYear, 5),
                                   range = TRUE)

popNewInfectionsPlot <- indicatorPlot(popInfections, 
                                   ylabel = "New infections",
                                   xlimits = c(startYear, endYear, 10),
                                   range = TRUE,
                                   facetPlot = "population")
# Cumulative incidence plots
tempTotalInfections <- indicatorResults(pg, bestResults, "newInfections",
                                        paramResults,
                                  range = FALSE,
                                  populations = "all", 
                                  annual = "sum") 
  
totalCumInfections <- tempTotalInfections %>%
  gather("sim", "infections", 2:(pg$parameter_samples + 2)) %>%
  group_by(sim) %>%
  mutate(cuminfections = cumsum(infections)) %>%
  select(-infections) %>%
  ungroup() %>%
  group_by(year) %>%
  summarise(min = min(cuminfections),
            max = max(cuminfections)) %>%
  ungroup() %>%
  mutate(best = cumsum(tempTotalInfections$best)) %>%
  select(year, best, min, max)
           
totalCumInfectionsPlot <- indicatorPlot(totalCumInfections, 
                                   ylabel = "Number of infections",
                                   xlimits = c(startYear, endYear, 5),
                                   range = TRUE)

tempPopInfections <- indicatorResults(pg, bestResults, 
                                      "newInfections", 
                                      paramResults,
                                      range = FALSE,
                                      annual = "sum") %>%
  FactorPop(popLabels)

bestEstimate <- tempPopInfections %>%
  select(year, population, best) %>%
  group_by(population) %>%
  mutate(cuminfects = cumsum(best)) %>%
  select(year, population, cuminfects)

popCumInfections <- tempPopInfections %>%
  gather("sim", "infections", 3:ncol(tempPopInfections)) %>%
  group_by(sim, population) %>%
  mutate(cuminfections = cumsum(infections)) %>%
  select(-infections) %>%
  ungroup() %>%
  group_by(year, population) %>%
  summarise(min = min(cuminfections),
            max = max(cuminfections)) %>%
  ungroup() %>%
  arrange(population) %>%
  mutate(best = bestEstimate$cuminfects) %>%
  select(year, population, best, min, max)

popCumInfectionsPlot <- indicatorPlot(popCumInfections, 
                                   ylabel = "Number of infections",
                                   xlimits = c(startYear, endYear, 10),
                                   range = TRUE,
                                   facetPlot = "population")

# Incidence --------------------------------------------------------------

tempTotalInfections <- select(tempTotalInfections, -year)
tempPopInfections <- select(tempPopInfections, -year)

tempTotalPop <- popResults(pg, bestResults, paramResults,
                       populations = "all", states = "all",
                       range = FALSE) %>% select(-year)
tempTotalPop <- tempTotalPop[MidyearIndex(nrow(tempTotalPop),
                            pg$timestep), ]


tempPopPop <- popResults(pg, bestResults, paramResults, states = "all",
                       range = FALSE) %>% 
  select(-year) %>%
  arrange(population) %>%
  FactorPop(popLabels)
tempPopPop <- tempPopPop[MidyearIndex(nrow(tempPopPop),
                            pg$timestep), ]

savePops <- tempPopPop$population
tempPopPop <- select(tempPopPop, -population)


#Calculate incidence per 100,000
tempTotalInc <- (1e5 * tempTotalInfections /
                   tempTotalPop) %>%
  mutate(year = head(pg$years, -1)) %>%
  select(year, everything()) %>%
  tbl_df()

totalInc <- tempTotalInc %>%
  gather("sim", "incidence", 2:ncol(tempTotalInc)) %>%
  group_by(year) %>%
  summarise(min = min(incidence),
            max = max(incidence)) %>%
  ungroup() %>%
  mutate(best = tempTotalInc$best) %>%
  select(year, best, min, max)

totalIncidencePlot <- indicatorPlot(totalInc, 
                               ylabel = "Incidence per 100,000",
                               xlimits = c(startYear, endYear, 5),
                               range = TRUE)

tempPopInc <- (1e5 * select(tempPopInfections, -population) / 
                 tempPopPop) %>%
  mutate(year = rep(head(pg$years, -1), pg$npops)) %>%
  select(year, everything()) %>%
  tbl_df() %>%
  mutate(population = savePops) %>%
  select(year, population, everything())

popInc <- tempPopInc %>%
  gather("sim", "incidence", 3:ncol(tempPopInc)) %>%
  group_by(year, population) %>%
  summarise(min = min(incidence),
            max = max(incidence)) %>%
  ungroup() %>%
  arrange(population) %>%
  mutate(best = tempPopInc$best) %>%
  select(year, population, best, min, max)

popIncidencePlot <- indicatorPlot(popInc, 
                               ylabel = "Incidence per 100,000",
                               xlimits = c(startYear, endYear, 10),
                               facetPlot = "population", 
                               range = TRUE)

# Treatment ------------------------------------------------------------ 

totalTreatments <- indicatorResults(pg, bestResults, "newTreatments", 
                                    paramResults,
                                    range = TRUE,
                                    populations = "all", 
                                    annual = "sum")

popTreatments <- indicatorResults(pg, bestResults, "newTreatments",
                                  paramResults,
                                  range = TRUE,
                                  annual = "sum")

totalTreatmentPlot <- indicatorPlot(totalTreatments, 
  ylabel = "Number initiated treatment", 
  range = TRUE, xlimits = c(startYear, endYear, 5))

popTreatmentPlot <- indicatorPlot(popTreatments, 
  ylabel = "Number initiated treatment", 
  range = TRUE, xlimits = c(startYear, endYear, 5),
  groupPlot = "population") +
  scale_colour_brewer(name = "Age Group", palette = "Set1",
                      labels = c("Age < 4", "Age 5 to 14", "Age 15 to 44",
                                "Age > 45")) +
  scale_fill_brewer(name = "Age Group", palette = "Set1",
                      labels = c("Age < 4", "Age 5 to 14", "Age 15 to 44",
                                "Age > 45"))

# HBV deaths -------------------------------------------------------------
totalDeaths <- indicatorResults(pg, bestResults, "newHBVdeaths",
                                paramResults,
                                range = TRUE,
                                populations = "all", 
                                annual = "sum")

popDeaths <- indicatorResults(pg, bestResults, 
                              "newHBVdeaths",
                              paramResults,
                              range = TRUE,
                              annual = "sum")

totalDeathsPlot <- indicatorPlot(totalDeaths, 
  ylabel = "Number of deaths", 
  range = TRUE, xlimits = c(startYear, endYear, 5))

popDeathsPlot <- indicatorPlot(popDeaths, 
  ylabel = "Number of deaths", 
  range = TRUE, xlimits = c(startYear, endYear, 5),
  groupPlot = "population") +
  scale_colour_brewer(name = "Age Group", palette = "Set1",
                      labels = c("Age < 4", "Age 5 to 14", "Age 15 to 44",
                                "Age > 45")) +
  scale_fill_brewer(name = "Age Group", palette = "Set1",
                      labels = c("Age < 4", "Age 5 to 14", "Age 15 to 44",
                                "Age > 45"))

# Vaccinations -----------------------------------------------------------


# Display plots (if required) --------------------------------------------

if (displayPlots) {
  totalPopPlot
  agePopPlot
  
  totalInfectedPlot
  totalPrevPlot
  popInfectedPlot
  popPrevPlot
  allPopsPlot
  allStatesPlot
  
  totalNewInfectionsPlot
  popNewInfectionsPlot
  totalCumInfectionsPlot
  popCumInfectionsPlot
  
  totalIncidencePlot
  popIncidencePlot
  
  totalTreatmentPlot
  popTreatmentPlot
  
  totalDeathsPlot
  popDeathsPlot
  
  # totalVaccinePlot
  # popVaccinePlot
}

# Save plots separately (if required) ------------------------------------
if (savePlots) {
  
  figureFolder <- file.path(projectFolder, project, "results", 
                         paste0("results_", resultsTime))
  
  SaveFigure(figureFolder, "total_population", totalPopPlot)
  SaveFigure(figureFolder, "age_group_population", agePopPlot)
  
  SaveFigure(figureFolder, "total_infected", totalInfectedPlot)
  SaveFigure(figureFolder, "total_prevalence", totalPrevPlot)
  SaveFigure(figureFolder, "age_group_infected", popInfectedPlot)
  SaveFigure(figureFolder, "age_group_prevalence", popPrevPlot)
  SaveFigure(figureFolder, "all_pops_states", allPopsPlot)
  SaveFigure(figureFolder, "all_states_pops", allStatesPlot)
  
  SaveFigure(figureFolder, "new_infections", totalNewInfectionsPlot)
  SaveFigure(figureFolder, "age_group_new_infections",
             popNewInfectionsPlot)
  SaveFigure(figureFolder, "cumulative_infections", totalCumInfectionsPlot)
  SaveFigure(figureFolder, "age_group_cumulative_infections",
             popCumInfectionsPlot)
  
  SaveFigure(figureFolder, "total_incidence", totalIncidencePlot)
  SaveFigure(figureFolder, "age_group_incidence", popIncidencePlot)
  
  SaveFigure(figureFolder, "total_treatment", totalTreatmentPlot)
  SaveFigure(figureFolder, "age_group_treatment", popTreatmentPlot)
  
  SaveFigure(figureFolder, "total_deaths", totalDeathsPlot)
  SaveFigure(figureFolder, "age_group_deaths", popDeathsPlot)
  
  # Vaccination
}
```

## Summary Figures

### Total population size

**Figure 1** - Total population size over time.

```{r Insert figure 1, include = TRUE}
totalPopPlot
```

**Figure 2** - Population size by age group over time.

```{r Insert figure 2, include = TRUE}
agePopPlot
```

### Prevalence

**Figure 3** - Total number of people infected with HBV over time.

```{r Insert figure 3, include = TRUE}
totalInfectedPlot
```

**Figure 4** - Overall population prevalence over time.

```{r Insert figure 4, include = TRUE}
totalPrevPlot
```

**Figure 5** - Number of people infected with HBV over time by age group.

```{r Insert figure 5, include = TRUE}
popInfectedPlot
```

**Figure 6** - HBV prevalence by age group over time.

```{r Insert figure 6, include = TRUE}
popPrevPlot
```


**Figure 7** - Number of people in each stage of infection by age_group.

```{r Insert figure 7, include = TRUE}
allPopsPlot
```

**Figure 8** - Number of people in each stage of infection by infection
state.

```{r Insert figure 8, include = TRUE}
allStatesPlot
```

### New infections and incidence

**Figure 9** - Total number of new HBV infections.

```{r Insert figure 9, include = TRUE}
totalNewInfectionsPlot
```

**Figure 10** - Number of new HBV infections by age group.

```{r Insert figure 10, include = TRUE}
popNewInfectionsPlot
```

**Figure 11** - Total number of new HBV infections.

```{r Insert figure 11, include = TRUE}
totalCumInfectionsPlot
```

**Figure 12** - Number of new HBV infections by age group.

```{r Insert figure 12, include = TRUE}
popCumInfectionsPlot
```


### Incidence

**Figure 13** - Overall incidence per 100,000 people.

```{r Insert figure 13, include = TRUE}
totalIncidencePlot
```

**Figure 14** - Incidence per 100,000 people by age group.

```{r Insert figure 14, include = TRUE}
popIncidencePlot
```

### HBV Treatment

**Figure 15** - Total number of people initiating treatment per year.

```{r Insert figure 15, include = TRUE}
totalTreatmentPlot
```

**Figure 16** - Number of people initiating treatment per year by age
group.

```{r Insert figure 16, include = TRUE}
popTreatmentPlot
```

### Deaths

**Figure 17** - Total number of deaths per year.

```{r Insert figure 17, include = TRUE}
totalDeathsPlot
```

**Figure 18** - Number of deaths per year by age group.

```{r Insert figure 18, include = TRUE}
popDeathsPlot
```

### Vaccination
